feat(#100): dedicating an own cli argument for external window recording
- introduce the new `-w | --win-id` flag
- update the docs
- refactor minor things
