+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "t-rec"
-version = "0.6.3"
+version = "0.7.0"
 authors = ["Sven Assmann <sven.assmann.it@gmail.com>"]
 edition = "2018"
 license = "GPL-3.0-only"
