+++ src/decor_effect.rs
@@ -5,7 +5,7 @@ use anyhow::Context;
 use rayon::prelude::*;
 use tempfile::TempDir;
 
-use crate::{file_name_for, Result};
+use crate::Result;
 
 ///
 /// apply a border decor effect via a chain of convert commands
@@ -100,7 +100,7 @@ fn apply_effect(
     effect: Box<dyn Fn(PathBuf) -> Result<()> + Send + Sync>,
 ) {
     time_codes.into_par_iter().for_each(|tc| {
-        let file = tempdir.path().join(file_name_for(tc, "tga"));
+        let file = tempdir.path().join(crate::utils::file_name_for(tc, "tga"));
         if let Err(e) = effect(file) {
             eprintln!("{}", e);
         }
