+++ src/linux/mod.rs
@@ -3,8 +3,8 @@ mod x11_api;
 use crate::{PlatformApi, Result};
 use x11_api::X11Api;
 
-pub fn setup() -> Result<Box<dyn PlatformApi>> {
-    Ok(Box::new(X11Api::new()?))
+pub fn setup() -> Result<impl PlatformApi> {
+    Ok(X11Api::new()?)
 }
 
 pub const DEFAULT_SHELL: &str = "/bin/sh";
